Module ckb_vm::machine

Modules

aot

asm

trace

Structs

DefaultCoreMachine

DefaultMachine

DefaultMachineBuilder

Traits

CoreMachine
This is the core part of RISC-V that only deals with data part, it is extracted from Machine so we can handle lifetime logic in dynamic syscall support.

Machine
This is the core trait describing a full RISC-V machine. Instruction package only needs to deal with the functions in this trait.

SupportMachine
This traits extend on top of CoreMachine by adding additional support such as ELF range, cycles which might be needed on Rust side of the logic, such as runner or syscall implementations.

Type Definitions

InstructionCycleFunc