Trait ckb_vm::machine::Machine

pub trait Machine: CoreMachine {
    fn ecall(&mut self) -> Result<(), Error>;

    fn ebreak(&mut self) -> Result<(), Error>;
}

This is the core trait describing a full RISC-V machine. Instruction package only needs to deal with the functions in this trait.

Required methods

fn ecall(&mut self) -> Result<(), Error>

fn ebreak(&mut self) -> Result<(), Error>

Loading content...