Trait ckb_vm::machine::SupportMachine

pub trait SupportMachine: CoreMachine {
    fn cycles(&self) -> u64;

    fn set_cycles(&mut self, cycles: u64);

    fn max_cycles(&self) -> Option<u64>;

    fn running(&self) -> bool;

    fn set_running(&mut self, running: bool);

    fn add_cycles(&mut self, cycles: u64) -> Result<(), Error> { ... }

    fn load_elf(
        &mut self, 
        program: &Bytes, 
        update_pc: bool
    ) -> Result<u64, Error> { ... }

    fn initialize_stack(
        &mut self, 
        args: &[Bytes], 
        stack_start: u64, 
        stack_size: u64
    ) -> Result<u64, Error> { ... }
}

This traits extend on top of CoreMachine by adding additional support such as ELF range, cycles which might be needed on Rust side of the logic, such as runner or syscall implementations.

Required methods

fn cycles(&self) -> u64

fn set_cycles(&mut self, cycles: u64)

fn max_cycles(&self) -> Option<u64>

fn running(&self) -> bool

fn set_running(&mut self, running: bool)

Loading content...
Provided methods

fn add_cycles(&mut self, cycles: u64) -> Result<(), Error>

fn load_elf(&mut self, program: &Bytes, update_pc: bool) -> Result<u64, Error>

fn initialize_stack(
&mut self,
args: &[Bytes],
stack_start: u64,
stack_size: u64
) -> Result<u64, Error>

Loading content...