Trait ckb_vm::machine::CoreMachine

pub trait CoreMachine {
    type REG: Register;
    type MEM: Memory<Self::REG>;
    fn pc(&self) -> &Self::REG;

    fn set_pc(&mut self, next_pc: Self::REG);

    fn memory(&self) -> &Self::MEM;

    fn memory_mut(&mut self) -> &mut Self::MEM;

    fn registers(&self) -> &[Self::REG];

    fn set_register(&mut self, idx: usize, value: Self::REG);
}

This is the core part of RISC-V that only deals with data part, it is extracted from Machine so we can handle lifetime logic in dynamic syscall support.

Associated Types

type REG: Register

type MEM: Memory<Self::REG>

Loading content...
Required methods

fn pc(&self) -> &Self::REG

fn set_pc(&mut self, next_pc: Self::REG)

fn memory(&self) -> &Self::MEM

fn memory_mut(&mut self) -> &mut Self::MEM

fn registers(&self) -> &[Self::REG]

fn set_register(&mut self, idx: usize, value: Self::REG)

Loading content...